feat: format addresses and tx hashes
