+++ src/features/messages/cards/TransactionCard.tsx
@@ -1,11 +1,5 @@
-import { ChainMetadata, MultiProvider } from '@hyperlane-xyz/sdk';
-import {
-  ProtocolType,
-  hexToRadixCustomPrefix,
-  isAddress,
-  isZeroish,
-  strip0x,
-} from '@hyperlane-xyz/utils';
+import { MultiProvider } from '@hyperlane-xyz/sdk';
+import { isAddress, isZeroish } from '@hyperlane-xyz/utils';
 import { Modal, SpinnerIcon, Tooltip, useModal } from '@hyperlane-xyz/widgets';
 import BigNumber from 'bignumber.js';
 import { PropsWithChildren, ReactNode, useState } from 'react';
@@ -14,6 +8,7 @@ import { Card } from '../../../components/layout/Card';
 import { links } from '../../../consts/links';
 import { useMultiProvider } from '../../../store';
 import { MessageStatus, MessageTx } from '../../../types';
+import { formatAddress, formatTxHash } from '../../../utils/addresses';
 import { getDateTimeString, getHumanReadableTimeString } from '../../../utils/time';
 import { ChainSearchModal } from '../../chains/ChainSearchModal';
 import { getChainDisplayName, isEvmChain } from '../../chains/utils';
@@ -193,17 +188,6 @@ function TransactionCard({
   );
 }
 
-function getFormattedTransactionHash(hash: string, metadata: ChainMetadata | null) {
-  switch (metadata?.protocol) {
-    case ProtocolType.Radix:
-      return hexToRadixCustomPrefix(hash, 'txid', metadata?.bech32Prefix);
-    case ProtocolType.Cosmos:
-      return strip0x(hash);
-    default:
-      return hash;
-  }
-}
-
 function TransactionDetails({
   chainName,
   domainId,
@@ -218,17 +202,10 @@ function TransactionDetails({
   blur: boolean;
 }) {
   const multiProvider = useMultiProvider();
-  const protocol = multiProvider.tryGetProtocol(domainId) || ProtocolType.Ethereum;
-  const metadata = multiProvider.tryGetChainMetadata(domainId);
-
   const { hash, from, timestamp, blockNumber, mailbox } = transaction;
 
-  const formattedHash = getFormattedTransactionHash(hash, metadata);
-
-  const formattedMailbox =
-    protocol === ProtocolType.Radix
-      ? hexToRadixCustomPrefix(mailbox, 'component', metadata?.bech32Prefix, 30)
-      : mailbox;
+  const formattedHash = formatTxHash(hash, domainId, multiProvider);
+  const formattedMailbox = formatAddress(mailbox, domainId, multiProvider);
 
   const txExplorerLink =
     hash && !new BigNumber(hash).isZero()
